🚚: split methods so that it can be copied over
